Generate unique combination of references
Will exhaust all possible combination of refs without retries.

Problems:

Creates the Cartesian product of refs, which can take a lot of memory
and time if you have a lot of refs. We may be able to optimize this.

The end result is a vector of refs as a single attribute value. In order
to use the refs in other parts you need to assign them to their own
attributes. See world_test.clj. Not sure what a redesigned system, which
allows for directly generating multiple attributes at the same time,
will look like.